⚡️ :: ui수정 & 컴포넌트 분리 & 자잘한 수정
