+++ src/assets/changeTeacher/index.tsx
@@ -2,6 +2,7 @@ interface IconType {
   direction: number;
   onClick?: () => void;
 }
+
 export default function CalendarArrowIcon({ direction, onClick }: IconType) {
   return (
     <svg
@@ -11,7 +12,7 @@ export default function CalendarArrowIcon({ direction, onClick }: IconType) {
       viewBox="0 0 12 20"
       fill="none"
       xmlns="http://www.w3.org/2000/svg"
-      style={{ transform: `rotate(${direction}deg)` }}
+      style={{ transform: `rotate(${direction}deg)`, cursor: "pointer" }}
     >
       <path
         d="M9.1258 19.1004L0.700797 10.7004C0.600797 10.6004 0.529797 10.4921 0.487797 10.3754C0.445797 10.2587 0.42513 10.1337 0.425797 10.0004C0.425797 9.86706 0.446797 9.74206 0.488797 9.62539C0.530797 9.50872 0.601464 9.40039 0.700797 9.30039L9.1258 0.875391C9.35913 0.642057 9.6508 0.525391 10.0008 0.525391C10.3508 0.525391 10.6508 0.650391 10.9008 0.900391C11.1508 1.15039 11.2758 1.44206 11.2758 1.77539C11.2758 2.10872 11.1508 2.40039 10.9008 2.65039L3.5508 10.0004L10.9008 17.3504C11.1341 17.5837 11.2508 17.8714 11.2508 18.2134C11.2508 18.5554 11.1258 18.8511 10.8758 19.1004C10.6258 19.3504 10.3341 19.4754 10.0008 19.4754C9.66746 19.4754 9.3758 19.3504 9.1258 19.1004Z"
