+++ src/components/club/member.tsx
@@ -2,8 +2,8 @@ import styled from "@emotion/styled";
 import Image from "next/image";
 import { useState } from "react";
 import { MoreIcon, Leader } from "@/assets/club";
-import Modal from "../common/modal";
-import DropDown from "../common/dropDown";
+import Modal from "../common/Modal";
+import DropDown from "../common/Dropdown";
 import Menu from "./menu";
 import { ItemType } from "@/models/common";
 import { layerDropDownItem } from "@/constants/DropDownItem";
